requirements.txt
fastapi
uvicorn
groq
pymupdf
docker
python-dotenv